Update use-directory-based-edge-blocking.md
This request is to add the following as a note: For a hybrid configuration, the recipient's SMTP domain must have the DNS MX record pointing to O365/EOP. This request came via CI 78407 and was approved by Sr. SPM Scott Landry. Please reword the sentence as appropriate for language/style. Thanks much!