Update terms
Microsoft Pinpoint appears to be renamed to Microsoft Solution Providers, at least that is where the URL leads to. So updating the terms. Also can this document have additional notes about how to search for fax solution providers under Microsoft Solution Providers as it does not appear to results. Or should another URL be referenced instead?