Updated steps for setting group writeback permissions
Updated the steps to configure the required permissions for the AAD Connect sync account. AAD Connect uses a separate account with an MSOL prefix to perform CRUD actions to the target AD forest. AAD Connect does not use the AAD_ service account. 
Added the $ADConnectorAccountDomain variable. because ADConnectorAccountDomain is a mandatory parameter.