Update use-hybrid-modern-auth.md
This section is not clear at all about some of the stuff it says.  For one, it's not required/necessary to even setup an Intune App Protection policy for Outlook mobile app to work with HMA.  Second, if choosing to optionally create an Intune App Protection policy for Outlook for iOS and Android, the policy only needs to include the Outlook app, not all the ones that are stated.  These false claims with no backing are the reason I'm suggesting to just delete the bad section altogether.