Create Disable or enable journaling of voice mail and missed call notifications
Please contact angelat if you have any questions. When this feature is disabled by setting the value to false all messages sent with the item class listed will by pass the Journaling Agent. It is important to highlight the messages are blocked by using the message class because a malicious actor could spoof messages with this item class to intentionally by pass the Journal Agent. Customer must be aware of this risk.