Update supportability-matrix.md
Shouldn't it say "UR" instead of "RU" since they are Update Rollups, not Rollup Updates? Really appreciate having this page - definitely one of my go-to pages and at the top of my bookmarks. Thanks for keeping it up.
See https://github.com/MicrosoftDocs/OfficeDocs-Exchange/pull/2254 for a similar change that was made on another doc.